refactor: modularize core responsibilities
